fixed stairs bug
walkmap now resizes all objects indiscriminately, instead of only objects which can't be stepped onto.  this fixes a bug where the player could collide with the edge of an inner step, at the cost of making it weird to look over ledges with an unproportionate player radius.  this should be unnoticeable with most reasonable values.